com.test.Personal 패키지 통합 및 관리자 그룹리스트 검색/필터링/정렬/페이징 완벽
