+++ MeetU2/src/com/test/mapper/IPublicDAO.xml
@@ -1,0 +1,170 @@
+﻿<?xml version="1.0" encoding="UTF-8" ?>
+<!DOCTYPE mapper
+ PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN"
+ "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
+<mapper namespace="com.test.dao.IPublicDAO">
+
+
+<!-- 관심사 리스트 -->
+	<select id="getCategorys" resultType="com.test.dto.OptionDTO">
+		SELECT CATEGORY_CODE AS VALUE, CATEGORY_CONTENT AS NAME
+		FROM TBL_CATEGORY
+		ORDER BY CATEGORY_CODE
+	</select>
+
+<!-- 지역 리스트 -->
+	<select id="getCity_types" resultType="com.test.dto.OptionDTO">
+		SELECT CITYPE_ID AS VALUE, CITY_NAME AS NAME
+		FROM CITY_TYPE
+		ORDER BY CITYPE_ID
+	</select>
+	
+<!-- 그룹 공개범위 리스트 -->
+	<select id="getGroupPublic" resultType="com.test.dto.OptionDTO">
+		SELECT GRPUBLIC_ID AS VALUE, GRPUBLIC_NAME AS NAME
+		FROM GRPUBLIC
+		ORDER BY GRPUBLIC_ID
+	</select>	
+	
+<!-- 시티타입코드 가져가서 이름가져오기 -->
+	<select id="getCity_name" resultType="String">
+		SELECT CITY_NAME AS NAME
+		FROM CITY_TYPE
+		WHERE CITYPE_ID = #{cityPe_id}
+	</select>
+
+<!-- 카테고리id 가져가서 이름가져오기 -->
+	<select id="getCategory_content" resultType="String">
+		SELECT CATEGORY_CONTENT AS NAME
+		FROM TBL_CATEGORY
+		WHERE CATEGORY_CODE = #{category_code}
+	</select>
+	
+<!-- 그룹공개id 가져가서 이름가져오기 -->
+	<select id="getGrPublic_name" resultType="String">
+		SELECT GRPUBLIC_NAME AS NAME
+		FROM GRPUBLIC
+		WHERE GRPUBLIC_ID = #{grPublic_id}
+	</select>
+
+
+
+
+	<select id="getGroupList" parameterType="com.test.dto.FilterDTO" resultType="com.test.admin.GroupDTO">
+		SELECT *
+		FROM(
+			SELECT ROWNUM AS ROWNUMBER, GG.* 
+			FROM(SELECT G.* 
+				 FROM(SELECT LGROUP_ID AS GROUP_ID, GROUP_NAME 
+					  , (SELECT LMEMBER_ID FROM GRLIST WHERE GRPOWER_ID='1' AND LGROUP_ID=T.LGROUP_ID) AS member_id 
+					  , (SELECT COUNT(*) FROM GRLIST WHERE LGROUP_ID=T.LGROUP_ID) AS group_memberCount 
+					  , TO_CHAR(GROUP_CREDATE, 'YYYY-MM-DD') AS group_credate 
+					  , (SELECT CITY_NAME FROM CITY_TYPE WHERE CITYPE_ID = T.CITYPE_ID) AS group_citypeName 
+					  , (SELECT LISTAGG(CATEGORY_CONTENT, '/') WITHIN GROUP(ORDER BY LGROUP_ID) 
+						 FROM VGRCAT WHERE LGROUP_ID=T.LGROUP_ID GROUP BY LGROUP_ID) AS group_category 
+					  , (SELECT GRPUBLIC_NAME FROM GRPUBLIC WHERE T.PUBLIC_GR = GRPUBLIC_ID) AS public_gr
+					  , (SELECT COUNT(*)FROM EVENT WHERE LGROUP_ID=T.LGROUP_ID AND SYSDATE>EVENT_DATE) AS PAST_EVENT
+					  , (SELECT COUNT(*)FROM EVENT WHERE LGROUP_ID=T.LGROUP_ID AND EVENT_DATE>SYSDATE) AS FUTURE_EVENT
+					  , (SELECT (case when (max(EVENT_CREDATE) is null) then '없음' else TO_CHAR(max(EVENT_CREDATE),'YYYY-MM-DD') end ) FROM EVENT WHERE LGROUP_ID=T.LGROUP_ID) AS CURRENT_CREATE
+	 			      , (SELECT (case when (max(EVENT_DATE) is null) then '없음' else TO_CHAR(max(EVENT_CREDATE),'YYYY-MM-DD') end ) FROM EVENT WHERE LGROUP_ID=T.LGROUP_ID) AS CURRENT_HOLD
+					  , (SELECT (CASE WHEN (SYSDATE-180>MAX(EVENT_CREDATE)   OR (SYSDATE-180>GROUP_CREDATE AND MAX(EVENT_CREDATE)IS NULL )) THEN '휴면' ELSE '활동' END)
+					     FROM EVENT WHERE LGROUP_ID=T.LGROUP_ID) AS GRSTATUS_STOP
+			          FROM TBL_GROUP T
+				  )G
+				  ORDER BY <choose>
+					 	<when test="sort == 1 || sort == 101">
+					 		GROUP_ID
+					 		<if test="sort == 1">
+					 			DESC
+					 		</if>
+					 	</when>
+					 	<when test="sort == 2 || sort == 102">
+					 		GROUP_NAME
+					 		<if test="sort == 2">
+					 			DESC
+					 		</if>
+					 	</when>
+					 	<when test="sort == 3 || sort == 103">
+					 		member_id
+					 		<if test="sort == 3">
+					 			DESC
+					 		</if>
+					 	</when>
+					 	<when test="sort == 4 || sort == 104">
+					 		group_memberCount
+					 		<if test="sort == 4">
+					 			DESC
+					 		</if>
+					 	</when>
+					 	<when test="sort == 5 || sort == 105">
+					 		group_credate
+					 		<if test="sort == 5">
+					 			DESC
+					 		</if>
+					 	</when>
+					 </choose>
+			) GG
+			<where>
+				<if test="group_name != null">
+			 		AND GROUP_NAME LIKE ('%' || #{group_name} || '%')
+			 	</if>
+			 	<if test="city_name != null">
+			 		AND group_citypeName LIKE ('%' || #{city_name} || '%')
+			 	</if>
+			 	<if test="category_content != null">
+			 		AND group_category LIKE ('%' || #{category_content} || '%')
+			 	</if>
+			 	<if test="grPublic_name != null">
+			 		AND public_gr LIKE ('%' || #{grPublic_name} || '%')
+			 	</if>
+			</where>
+			
+		)
+		 <where>
+		    AND ROWNUMBER BETWEEN #{startNum} AND #{endNum}  
+		 	
+		 </where>
+		 
+	</select>
+	
+	<select id="getGroupCount" parameterType="com.test.dto.FilterDTO" resultType="int">
+		SELECT COUNT(*)
+		FROM(
+			SELECT ROWNUM AS ROWNUMBER, GG.* 
+			FROM(SELECT G.* 
+				 FROM(SELECT LGROUP_ID AS GROUP_ID, GROUP_NAME 
+					  , (SELECT LMEMBER_ID FROM GRLIST WHERE GRPOWER_ID='1' AND LGROUP_ID=T.LGROUP_ID) AS member_id 
+					  , (SELECT COUNT(*) FROM GRLIST WHERE LGROUP_ID=T.LGROUP_ID) AS group_memberCount 
+					  , TO_CHAR(GROUP_CREDATE, 'YYYY-MM-DD') AS group_credate 
+					  , (SELECT CITY_NAME FROM CITY_TYPE WHERE CITYPE_ID = T.CITYPE_ID) AS group_citypeName 
+					  , (SELECT LISTAGG(CATEGORY_CONTENT, '/') WITHIN GROUP(ORDER BY LGROUP_ID) 
+						 FROM VGRCAT WHERE LGROUP_ID=T.LGROUP_ID GROUP BY LGROUP_ID) AS group_category 
+					  , (SELECT GRPUBLIC_NAME FROM GRPUBLIC WHERE T.PUBLIC_GR = GRPUBLIC_ID) AS public_gr
+					  , (SELECT COUNT(*)FROM EVENT WHERE LGROUP_ID=T.LGROUP_ID AND SYSDATE>EVENT_DATE) AS PAST_EVENT
+					  , (SELECT COUNT(*)FROM EVENT WHERE LGROUP_ID=T.LGROUP_ID AND EVENT_DATE>SYSDATE) AS FUTURE_EVENT
+					  , (SELECT (case when (max(EVENT_CREDATE) is null) then '없음' else TO_CHAR(max(EVENT_CREDATE),'YYYY-MM-DD') end ) FROM EVENT WHERE LGROUP_ID=T.LGROUP_ID) AS CURRENT_CREATE
+	 			      , (SELECT (case when (max(EVENT_DATE) is null) then '없음' else TO_CHAR(max(EVENT_CREDATE),'YYYY-MM-DD') end ) FROM EVENT WHERE LGROUP_ID=T.LGROUP_ID) AS CURRENT_HOLD
+					  , (SELECT (CASE WHEN (SYSDATE-180>MAX(EVENT_CREDATE)   OR (SYSDATE-180>GROUP_CREDATE AND MAX(EVENT_CREDATE)IS NULL )) THEN '휴면' ELSE '활동' END)
+					     FROM EVENT WHERE LGROUP_ID=T.LGROUP_ID) AS GRSTATUS_STOP
+			          FROM TBL_GROUP T
+				  )G
+			) GG
+			<where>
+				<if test="group_name != null">
+			 		AND GROUP_NAME LIKE ('%' || #{group_name} || '%')
+			 	</if>
+			 	<if test="city_name != null">
+			 		AND group_citypeName LIKE ('%' || #{city_name} || '%')
+			 	</if>
+			 	<if test="category_content != null">
+			 		AND group_category LIKE ('%' || #{category_content} || '%')
+			 	</if>
+			 	<if test="grPublic_name != null">
+			 		AND public_gr LIKE ('%' || #{grPublic_name} || '%')
+			 	</if>
+			</where>
+			
+		)
+		 
+	</select>
+</mapper>
