--- MeetU2/src/com/test/mapper/PersonalGroupDAO.xml
@@ -1,70 +1,0 @@
-<?xml version="1.0" encoding="UTF-8" ?>
-<!DOCTYPE mapper
- PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN"
- "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
-<mapper namespace="com.test.personal.IPersonalGroupDAO">
-
-
-<!-- 페이지 주인이 회원인 그룹리스트(그룹이 전체공개인 경우 수정이 가능) -->
-	<select id="groupList" resultType="com.test.personal.PersonalGroupDTO">
-		SELECT G.LGROUP_ID AS LGROUP_ID, G.LMEMBER_ID AS LMEMBER_ID, G.MBPUBLICTI_ID AS MBPUBLICTI_ID, TG.GROUP_NAME AS GROUP_NAME, GF.GRPROFILE_URL AS GRPROFILE_URL, G.GRPOWER_ID AS GRPOWER_ID
-		FROM GRLIST G JOIN TBL_GROUP TG 
-		ON G.LGROUP_ID = TG.LGROUP_ID
-        JOIN GRPROFILE GF
-        ON G.LGROUP_ID = GF.LGROUP_ID
-		WHERE G.LMEMBER_ID = #{userNumber } AND G.GRPOWER_ID IN(2,3) AND TG.PUBLIC_GR = 1
-	</select>
-	
-<!-- 페이지 주인이 회원인 그룹리스트(그룹이 멤버공개인 경우 수정이 불가)  -->	
-    <select id="groupList_secret" resultType="com.test.personal.PersonalGroupDTO">
-		SELECT G.LGROUP_ID AS LGROUP_ID, G.LMEMBER_ID AS LMEMBER_ID, G.MBPUBLICTI_ID AS MBPUBLICTI_ID, TG.GROUP_NAME AS GROUP_NAME, GF.GRPROFILE_URL AS GRPROFILE_URL, G.GRPOWER_ID AS GRPOWER_ID
-		FROM GRLIST G JOIN TBL_GROUP TG 
-		ON G.LGROUP_ID = TG.LGROUP_ID
-        JOIN GRPROFILE GF
-        ON G.LGROUP_ID = GF.LGROUP_ID
-		WHERE G.LMEMBER_ID = #{userNumber } AND G.GRPOWER_ID IN(2,3) AND TG.PUBLIC_GR = 3
-	</select>
-
-	
-<!-- 페이지 주인이 그룹장인 그룹리스트(그룹이 전체공개인 경우 수정이 가능)  -->	
-	<select id="groupList_owner" resultType="com.test.personal.PersonalGroupDTO">
-		SELECT G.LGROUP_ID AS LGROUP_ID, G.LMEMBER_ID AS LMEMBER_ID, G.MBPUBLICTI_ID AS MBPUBLICTI_ID, TG.GROUP_NAME AS GROUP_NAME, GF.GRPROFILE_URL AS GRPROFILE_URL, G.GRPOWER_ID AS GRPOWER_ID
-		FROM GRLIST G JOIN TBL_GROUP TG 
-		ON G.LGROUP_ID = TG.LGROUP_ID
-        JOIN GRPROFILE GF
-        ON G.LGROUP_ID = GF.LGROUP_ID
-		WHERE G.LMEMBER_ID = #{userNumber } AND G.GRPOWER_ID = 1 AND TG.PUBLIC_GR = 1
-	</select>
-	
-<!-- 페이지 주인이 그룹장인 그룹리스트(그룹이 멤버공개인 경우 수정이 불가)  -->	
-	<select id="groupList_owner_secret" resultType="com.test.personal.PersonalGroupDTO">
-		SELECT G.LGROUP_ID AS LGROUP_ID, G.LMEMBER_ID AS LMEMBER_ID, G.MBPUBLICTI_ID AS MBPUBLICTI_ID, TG.GROUP_NAME AS GROUP_NAME, GF.GRPROFILE_URL AS GRPROFILE_URL, G.GRPOWER_ID AS GRPOWER_ID
-		FROM GRLIST G JOIN TBL_GROUP TG 
-		ON G.LGROUP_ID = TG.LGROUP_ID
-        JOIN GRPROFILE GF
-        ON G.LGROUP_ID = GF.LGROUP_ID
-		WHERE G.LMEMBER_ID = #{userNumber } AND G.GRPOWER_ID = 1 AND TG.PUBLIC_GR = 3
-	</select>	
-
-<!-- 그룹리스트에서 뿌려준 그룹 탈퇴하기   -->	
-	<update id="groupOut">
-		UPDATE GRLIST
-      	SET GRPOWER_ID = 5 , GRLEAVE_DAY = SYSDATE
-      	WHERE LGROUP_ID = #{groupid } AND LMEMBER_ID = #{keynumber }
-	</update>
-	
-<!-- 그룹리스트에서 뿌려준 그룹 공개여부 비공개로 -->
-   	<update id="groupNoopen">
-   		UPDATE GRLIST
-		SET MBPUBLICTI_ID =5
-		WHERE LGROUP_ID = #{groupid } AND LMEMBER_ID = #{keynumber }
-   	</update>
-   	
-<!-- 그룹리스트에서 뿌려준 그룹 공개여부 공개로 -->   	
-	<update id="groupOpen">
-   		UPDATE GRLIST
-		SET MBPUBLICTI_ID = 1
-		WHERE LGROUP_ID = #{groupid } AND LMEMBER_ID = #{keynumber }
-   	</update>
-	
-</mapper>
