--- MeetU2/src/com/test/mapper/PersonalHomeDAO.xml
@@ -1,40 +1,0 @@
-<?xml version="1.0" encoding="UTF-8" ?>
-<!DOCTYPE mapper
- PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN"
- "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
-<mapper namespace="com.test.personal.IPersonalHomeDAO">
-	
-	
-<!-- 개인 소개 -->	
-<select id="userContent" resultType="String">
-	SELECT MEMBER_INFO AS USERCONTENT
-	FROM MU_MEMBER
-	WHERE LMEMBER_ID = #{userNumber}
-</select>
-
-
-<!-- 내가 주최한 이벤트 -->
-<select id="userEvent" resultType="com.test.personal.PersonalHomeDTO">
-	SELECT EVENT_NAME AS EVENTNAME, LOAD_NAME AS ADDRESS, EVENT_PLACE AS MAINADDR
-	     , TO_CHAR(EVENT_DATE, 'MM"월" DD"일" DAY') AS EVENTDATE1, TO_CHAR(EVENT_DATE, 'HH24 : MI') AS EVENTDATE2
-	     , COUNT AS GUESTCOUNT, EVENT_MAXATTEND AS MAXCOUNT
-	FROM VIEW_EVENT
-	WHERE LMEMBER_ID = #{userNumber}
-	ORDER BY EVENT_DATE DESC
-</select>
-
-	
-<!-- 참가 예정 이벤트 -->
-<select id="goEvent" resultType="com.test.personal.PersonalHomeDTO">
-	SELECT EVENT_NAME AS EVENTNAME, LOAD_NAME AS ADDRESS, EVENT_PLACE AS MAINADDR
-	     , TO_CHAR(EVENT_DATE, 'MM"월" DD"일" DAY') AS EVENTDATE1, TO_CHAR(EVENT_DATE, 'HH24 : MI') AS EVENTDATE2
-	     , COUNT AS GUESTCOUNT, EVENT_MAXATTEND AS MAXCOUNT
-	FROM VIEW_ATTENDEVENT
-	WHERE EVENT_DATE >= SYSDATE
-	AND ATTEND_ID = 1
-	AND GUEST = #{userNumber}
-	ORDER BY EVENT_DATE DESC
-</select>
-
-
-</mapper>
