--- MeetU2/src/com/test/mapper/PersonalLeftbarDAO.xml
@@ -1,74 +1,0 @@
-<?xml version="1.0" encoding="UTF-8" ?>
-<!DOCTYPE mapper
- PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN"
- "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
-<mapper namespace="com.test.personal.IPersonalLeftbarDAO">
-	
-<!-- 나의 프로필 변경하기  -->
-<update id="profilePhotoModify">
- update mbprofile
- set MBPROFILE_URL =  #{url}
- where  LMEMBER_ID= #{keynumber}
-</update>
-
-	
-<!-- 개인 프로필 사진 -->	
-<select id="profilePhoto" resultType="String">
-	SELECT MBPROFILE_URL AS PROFILEPHOTO
-	FROM MBPROFILE
-	WHERE LMEMBER_ID = #{userNumber}
-</select>
-
-	
-<!-- 이름 -->
-<select id="memberName" resultType="String">
-	SELECT MEMBER_NAME AS NAME
-	FROM MU_MEMBER
-	WHERE LMEMBER_ID = #{userNumber}
-</select>
-
-
-<!-- 친구 수 -->
-<select id="friendCount" resultType="int">
-	SELECT COUNT(*)	AS FRIENDCOUNT
-	FROM FRIEND	
-	WHERE FRSTATUS_ID = 2
-	  AND LIMEMBER_ID = #{userNumber}
-	   OR LUMEMBER_ID = #{userNumber}
-</select>
-
-
-<!-- 과거 meet U 수 -->
-<select id="pastMeetU" resultType="int">
-	SELECT NVL(COUNT(*),0) AS PASTMEETCOUNT	
-	FROM EVEATTEND			
-	WHERE LMEMBER_ID = #{userNumber}		
-	  AND ATTEND_ID = 1
-	  AND LEVENT_ID = ( SELECT LEVENT_ID 
-	                    FROM EVENT 
-	                    WHERE EVENT.LEVENT_ID = EVEATTEND.LEVENT_ID 
-	                    AND EVENT.EVENT_DATE <![CDATA[ < ]]> SYSDATE )
-</select>
-
-
-<!-- 관심 지역 -->
-<select id="userCityList" resultType="com.test.personal.PersonalLeftbarDTO">
-	SELECT CITY_NAME AS USERCITY
-	FROM MU_MEMBER M JOIN CITY_TYPE C
-	ON M.CITYPE_ID = C.CITYPE_ID
-	WHERE M.LMEMBER_ID = #{userNumber}
-</select>
-
-
-<!-- 관심사 -->
-<select id="userTagList" resultType="com.test.personal.PersonalLeftbarDTO">
-	SELECT C.CATEGORY_CONTENT AS USERTAG
-	FROM TBL_CATEGORY C JOIN MBCATEGORY MC
-	ON C.CATEGORY_CODE = MC.CATEGORY_CODE
-	WHERE MC.LMEMBER_ID = #{userNumber}
-</select>
-
-
-
-
-</mapper>
