--- MeetU2/src/com/test/personal/IPersonalPageIntroducemyselfDAO.java
@@ -1,8 +1,0 @@
-package com.test.personal;
-
-public interface IPersonalPageIntroducemyselfDAO
-{
-	// 개인 정보
-	public PersonalPageIntroducemyselfDTO userInfo(String userNumber);
-	
-}
